updates for segmentation
